pubsub: separate commands
